// res/interaction.d.ts
// Generated by https://quicktype.io

export interface Interaction {
    type:       number;
    token:      string;
    member:     Member;
    id:         string;
    guild_id:   string;
    data:       Data;
    channel_id: string;
}

export interface Data {
    options: Option[];
    name:    string;
    id:      string;
}

export interface Option {
    name:  string;
    value: string;
}

export interface Member {
    user:          User;
    roles:         string[];
    premium_since: null;
    permissions:   string;
    pending:       boolean;
    nick:          null;
    mute:          boolean;
    joined_at:     string;
    is_pending:    boolean;
    deaf:          boolean;
}

export interface User {
    id:            number;
    username:      string;
    avatar:        string;
    discriminator: string;
    public_flags:  number;
}