// res/ready.d.ts
export interface Ready {
    v: number;
    user: User;
    guilds: [Unavailable_Guilds];

    session_id: string;

    shard?: [Shard];
    
    application: Application
}

export interface User {
    // Generated by https://quicktype.io
    id:            string;
    username:      string;
    discriminator: string;
    avatar:        string;
    verified:      boolean;
    email:         string;
    flags:         number;
    premium_type:  number;
    public_flags:  number;
}

export interface Unavailable_Guilds {
    id:          string;
    unavailable: boolean;
}

export interface Shard {
    shard_id:   number;
    num_shards: number;
}

export interface Application {
    // Discord API doesn't have a reference
    id:    any;
    flags: any;
}